+++ Src/Private/Get-AbrAzSubscription.ps1
@@ -5,7 +5,7 @@ function Get-AbrAzSubscription {
     .DESCRIPTION
 
     .NOTES
-        Version:        0.1.0
+        Version:        0.1.1
         Author:         Tim Carman
         Twitter:        @tpcarman
         Github:         tpcarman
@@ -24,6 +24,10 @@ function Get-AbrAzSubscription {
 
     process {
         if ($AzSubscriptions) {
+            if ($Options.ShowSectionInfo) {
+                Paragraph "An Azure subscription is a logical container used to provision resources in Azure. It holds the details of all your resources like virtual machines (VMs), databases, and more. When you create an Azure resource like a VM, you must identify the subscription it belongs to."
+                BlankLine
+            }
             Paragraph "The following table summarises the subscription information within the $($AzTenant.Name) tenant."
             BlankLine
             $AzSubscriptionInfo = @()
@@ -46,7 +50,7 @@ function Get-AbrAzSubscription {
             }
             $AzSubscriptionInfo | Table @TableParams
         } else {
-            Write-PScriboMessage 'No subscriptions found.'
+            Write-PScriboMessage -IsWarning 'No subscriptions found.'
             Break
         }
     }